Return annotation on temporary home.work folder in TestPluginRDFRankBase.java to be  @Rule. Add temporary home.work folder to TestPluginRDFRankWithRDFStarData.java and TestPluginRDFRank.java
