+++ src/test/java/com/ontotext/trree/plugin/rdfrank/TestPluginRDFRankBase.java
@@ -38,8 +38,8 @@ public class TestPluginRDFRankBase {
 
 	protected boolean useUpdate;
 
-	@ClassRule
-	public static TemporaryLocalFolder tmpFolder = new TemporaryLocalFolder();
+	@Rule
+	public TemporaryLocalFolder tmpFolder = new TemporaryLocalFolder();
 
 	public TestPluginRDFRankBase(boolean useUpdate) {
 		this.useUpdate = useUpdate;
@@ -48,18 +48,6 @@ public class TestPluginRDFRankBase {
 	@Parameterized.Parameters
 	public static List<Object[]> getParameters() {
 		return Arrays.<Object[]> asList(new Object[] { true }, new Object[] { false });
-	}
-
-	@BeforeClass
-	public static void setWorkDir() {
-		System.setProperty("graphdb.home.work", String.valueOf(tmpFolder.getRoot()));
-		Config.reset();
-	}
-
-	@AfterClass
-	public static void resetWorkDir() {
-		System.clearProperty("graphdb.home.work");
-		Config.reset();
 	}
 
 	@Before
